Treewide: Clean every warnings
* Clang 17

+++ drivers/nfc/oplus_nfc/oplus_nfc.c
@@ -72,7 +72,7 @@ static int nfc_read_func(struct seq_file *s, void *v)
 {
 	void *p = s->private;
 
-	switch((uint32_t)(p)) {
+	switch((uintptr_t)(p)) {
 	case NFC_CHIPSET_VERSION:
 		seq_printf(s, "%s", current_chipset);
 		break;
